Feature/users

+++ app/Http/Controllers/EventController.php
@@ -4,7 +4,9 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Models\Fileentry;
 use App\Models\Oldslug;
+use App\Models\User;
 use App\Models\Type;
+use App\Models\UserEvent;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Event;
@@ -33,8 +35,8 @@ class EventController extends CommonController {
     {
         // Middleware définissant les pages où l'on ne peut accéder uniquement si l'on est connecté
         $this->middleware('auth', ['only' => ['create', 'edit', 'destroy']]);
+        // Middleware permettant d'effectuer les redirections 301
         $this->middleware('oldslug', ['only' => ['show', 'edit']]);
-        //$this->middleware('admin');
     }
 
 
@@ -87,7 +89,11 @@ class EventController extends CommonController {
      */
     public function show(Event $event)
     {
-        return view('events.show', compact('event'));
+        $type = Type::find($event->type_id);
+        $author = User::find($event->user_id);
+        $went = $this->userWent(Auth::user(), $event);
+
+        return view('events.show', compact('event', 'type', 'author', 'went'));
     }
 
     /**
@@ -110,25 +116,29 @@ class EventController extends CommonController {
      */
     public function update(Event $event)
     {
-        // TODO check si ce n'est pas une bêtise de retirer le token ?
-        $input = array_except(Input::all(), ['_token', '_method', '_wysihtml5_mode']);
+        if(Auth::user()->allowed('edit.event', $event)) {
+            // TODO check si ce n'est pas une bêtise de retirer le token ?
+            $input = array_except(Input::all(), ['_token', '_method', '_wysihtml5_mode']);
 
-        /* Saving old slug for 301 redirections */
-        if($input['name'] != $event->name){
-            $oldslug = Oldslug::create([
-                'event_id' => $event->id,
-                'slug' => $event->slug
-            ]);
-            $oldslug->save();
+            /* Saving old slug for 301 redirections */
+            if ($input['name'] != $event->name) {
+                $oldslug = Oldslug::create([
+                    'event_id' => $event->id,
+                    'slug' => $event->slug
+                ]);
+                $oldslug->save();
+            }
+
+            /* If a file is sent */
+            if (Rqst::file()) {
+                $input['poster'] = $this->imageUpload('poster', true);
+            }
+            $event->update($input);
+
+            return Redirect::route('event.show', [$event->slug])->with('message', 'Évènement modifié');
+        } else {
+            return Redirect::route('event.index')->with('message', 'Vous n\'avez pas les permissions requises');
         }
-
-        /* If a file is sent */
-        if(Rqst::file()) {
-            $input['poster'] = $this->imageUpload('poster', true);
-        }
-        $event->update($input);
-
-        return Redirect::route('event.show', [$event->slug])->with('message', 'Évènement modifié');
     }
 
     /**
@@ -139,8 +149,63 @@ class EventController extends CommonController {
      */
     public function destroy(Event $event)
     {
-        $event->delete();
-        return Redirect::route('event.index')->with('message', 'Évènement supprimé');
+        // On vérifie, au cas où, si l'utilisateur possède bien les droits de suppression de l'évènement.
+        if(Auth::user()->allowed('delete.event', $event)) {
+
+            $oldslugs = Oldslug::where('event_id', $event->id);
+
+            $event->delete();
+
+            $oldslugs->delete();
+
+            return Redirect::route('event.index')->with('message', 'Évènement supprimé');
+        } else {
+            return Redirect::route('event.index')->with('message', 'Vous n\'avez pas les permissions requises');
+        }
+
+    }
+
+    /**
+     * Checks if a user is going or went to a specific event
+     *
+     * @param User $user
+     */
+    public function userWent(User $user, Event $event) {
+
+        $went = UserEvent::where('user_id', $user->id)
+            ->where('event_id', $event->id)
+            ->get();
+
+        if(!empty($went[0])){
+            return true;
+        }
+
+        return false;
+    }
+
+
+    /**
+     * Add/Remove user to the event
+     *
+     * @param $event_id
+     * @return mixed
+     */
+    public function userGoing($event_id)
+    {
+
+        $user = Auth::user();
+        $event = Event::find($event_id);
+
+        // TODO currently checked twice in process, see if it can be improved
+        $went = $this->userWent($user, $event);
+
+        if($went){
+            $event->users()->detach($user->id);
+        } else {
+            $event->users()->attach($user->id);
+        }
+
+        return Redirect::route('event.show', [$event->slug]);
     }
 
 }
